requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
python-dotenv
pydantic_settings
sqlalchemy